[C++] Add currentChunkStart() / currentChunkEnd() to McapWriter (#1)
### Changelog
Adds two getter methods to the C++ `McapWriter` class to expose the
in-progress chunk's start/end timestamps.

### Docs

None

### Description

The C++ McapWriter class currently supports one built-in rule for
automatic chunking which is based on the sum of uncompressed message
size. In addition to (or in lieu of) this rule, it is often useful to
split chunks based on a fixed timing cadence such as every N seconds.
This PR exposes `McapWriter::currentChunkStart()` and
`McapWriter::currentChunkEnd()` which can be used for that calculation.